change structure types
