+++ src/types/utils.ts
@@ -1,0 +1,54 @@
+import { NestedValue } from './types';
+
+export type Primitive = string | boolean | number | symbol | null | undefined;
+
+export type EmptyObject = { [K in string | number]: never };
+
+export type NonUndefined<T> = T extends undefined ? never : T;
+
+export type LiteralToPrimitive<T extends any> = T extends string
+  ? string
+  : T extends number
+  ? number
+  : T extends boolean
+  ? boolean
+  : T;
+
+export type DeepPartial<T> = {
+  [K in keyof T]?: T[K] extends Array<infer U>
+    ? Array<DeepPartial<U>>
+    : T[K] extends ReadonlyArray<infer U>
+    ? ReadonlyArray<DeepPartial<U>>
+    : T[K] extends Record<string, unknown>
+    ? DeepPartial<T[K]>
+    : T[K];
+};
+
+export type DeepMap<T, TValue> = {
+  [K in keyof T]?: IsAny<T[K]> extends true
+    ? any
+    : NonUndefined<T[K]> extends NestedValue | Date | FileList
+    ? TValue
+    : NonUndefined<T[K]> extends object
+    ? DeepMap<T[K], TValue>
+    : NonUndefined<T[K]> extends Array<infer U>
+    ? IsAny<U> extends true
+      ? Array<any>
+      : U extends NestedValue | Date | FileList
+      ? Array<TValue>
+      : U extends object
+      ? Array<DeepMap<U, TValue>>
+      : Array<TValue>
+    : TValue;
+};
+
+export type IsAny<T> = boolean extends (T extends never ? true : false)
+  ? true
+  : false;
+
+export type IsFlatObject<T extends object> = Extract<
+  Exclude<T[keyof T], NestedValue | Date | FileList>,
+  any[] | object
+> extends never
+  ? true
+  : false;
