+++ src/utils/isEmptyObject.ts
@@ -1,5 +1,5 @@
 import isObject from './isObject';
-import { EmptyObject } from '../types/types';
+import { EmptyObject } from '../types/utils';
 
 export default (value: unknown): value is EmptyObject =>
   isObject(value) && !Object.keys(value).length;
