+++ src/utils/isPrimitive.ts
@@ -1,6 +1,6 @@
 import isNullOrUndefined from './isNullOrUndefined';
 import { isObjectType } from './isObject';
-import { Primitive } from '../types/types';
+import { Primitive } from '../types/utils';
 
 export default (value: unknown): value is Primitive =>
   isNullOrUndefined(value) || !isObjectType(value);
